Rename metadata.rs to info.rs and move warm_up_workers to config.rs
metadata.rs contained runtime query functions (versions, themes, locales,
worker stats, JS bundling) plus warm_up_workers which is an action, not
a query. Move warm_up_workers + its asyncio variant to config.rs where
the other converter setup functions live, then rename the remaining file
to info.rs which better describes its contents: pure informational
queries about the runtime environment.
